Moved utils into separate file
